Actualización de la configuración de la API y eliminación de repositorios y casos de uso de inicio de sesión obsoletos. Reorganización de importaciones y creación de barriles para widgets y enumeraciones.
