+++ lib/src/Core/Barrels/enums_barrel.dart
@@ -1,0 +1,1 @@
+export '../../Utils/Enums/list_color.dart';
