+++ lib/src/Core/Barrels/widgets_shared_barrel.dart
@@ -1,0 +1,6 @@
+export '../../Shared/Widgets/alert_dialog.dart';
+export '../../Shared/Widgets/app_bar.dart';
+export '../../Shared/Widgets/banner.dart';
+export '../../Shared/Widgets/button.dart';
+export '../../Shared/Widgets/card.dart';
+export '../../Shared/Widgets/text_field.dart';
