--- lib/src/Features/Login/Domain/Repositories/login_repository.dart
@@ -1,5 +1,0 @@
-import '../Entities/user_entity.dart';
-
-abstract class LoginRepository {
-  Future<UserEntity> login(final String username);
-}
