+++ lib/src/Shared/Widgets/text_field.dart
@@ -8,7 +8,7 @@ library;
 
 import 'package:flutter/material.dart';
 
-import '../../Utils/Enums/list_color.dart';
+import '../../Core/Barrels/enums_barrel.dart';
 
 /// Widget de campo de texto reutilizable.
 class CustomTextField extends StatelessWidget {
